Obey :filter_html restriction
Added in the accessors for security restrictions from the previous version of RedCloth.  With this commit, only adding :fliter_html support, but the rest is to come.  Had to do a lot of refactoring to get this to work, but it paves the way for cleaning up some older kluges.