Refactor HTML entity escaping.
Eliminates the options hash that was used to enable escaping for HTML but disable it for LaTeX. Implemented Gaspard Bucher's suggestion (http://www.ruby-forum.com/topic/143548#644326) of hooks to do the escaping.  The escaping itself is still in C, but it is called through the formatter class, so each formatter can specify which function to use for escaping.