Fix uninitialized constant RedCloth::TextileDoc
The problem was the workaround introduced to deal with bug #30 would remove not only the old RedCloth when it was loaded by Rails' textilize helper because the current RedCloth was frozen and not installed, but also the current version when it was installed. Rather sloppy on my part. [#42 state:resolved]