fix bug related to the command line consumption of the optional port
This means you can no longer specify the port, which is annoying,
but it doesn't break backwards compatibility with the common
redcloth as a translator usage. I don't play with enough ruby to
know if optparse can backtrack and I don't have time to
make it backtrack, so for now, this will have to do.