Stop extending into ERB the `t` and `textilize`
Prevent injection of the `t` and `textilize` methods into a `ERB` module through the `Util` module. When using the `i18n` gem in Rails it uses textilize method, instead of `translate` (aliased as `t`) helper. It will be still possible to inject textilize into a ERB after this commit with a manual code:

```
begin
  require 'erb'
  require 'redcloth/erb_extension'
  include ERB::Util
rescue LoadError
end
```