Fix loading problems when Gem is built on Windows
On Windows, the loading code in redcloth.rb assumed that one is loading it from a 'fat binary' gem. Using standard windows Ruby versions from 2.4 forward, the gem can be compiled by the user.

This change first tries to require the library using the prefix (for fat gems). If that fails, it tries again to require without the prefix (for nuilt gems).

Fixes #51 and #61.